+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Fixed
 - Fixed $id embed issue (Fixes #353)
 - Allow all Parts to be null as well as an empty string (Fixes #349)
+- Fix issue placing marker along coastline when using Mapbox (Fixes #322)
 
 ## 4.0.2 - 2022-06-08
 ### Fixed